MFT-1150

+++ i21Database/dbo/Stored Procedures/uspTFGenerateGT103.sql
@@ -35,7 +35,7 @@ BEGIN TRY
 
 	IF @Refresh = 1
 	BEGIN
-		DELETE FROM tblTFTransactionSummary
+		DELETE FROM tblTFTransactionSummary  WHERE intTaxAuthorityId = 14 AND strFormCode = 'GT-103'
 	END
 
 	-- ======================== HEADER ==============================
